refactor: split discipline_registry.py into focused modules
Extract DisciplineRule dataclass, conf parsing, and conf loading into
separate internal modules. Original file preserved as re-export entry point.
